requirements.txt
Flask==0.8
Flask-WTF==0.6
Jinja2==2.6
Mako==0.7.3
MarkupSafe==0.15
MySQL-python==1.2.4
SQLAlchemy==0.7.7
WTForms==1.0.1
Werkzeug==0.8.3
alembic==0.4.1
bencode==1.0
## FIXME: could not find svn URL in dependency_links for this package:
flup==1.0.3.dev-20110405
lockfile==0.9.1
pyftpdlib==0.7.0
pysendfile==2.0.0
python-daemon==1.6
wsgiref==0.1.2
